q2

+++ q2.py
@@ -31,7 +31,7 @@ def main(argv):
     with open(argv[2], 'w') as out:
         for i in output:
 
-            out.write('%s,\t%s\n' % (i[0], int(i[1][0])/i[1][1]))
+            out.write('%s,\t%s\n' % (i[0], float(i[1][0])/i[1][1]))
 
 
 if __name__ == '__main__':
